Add files via upload
Replace your database with the "sql" file I uploaded. I had to change a few things to get stuff to work. Honestly, we just need a search bar and filter, constraints for appointments (none should show up that have a past date in the search appointment page), medical history upload with prescriptions for each and pdf printout of medical history. 

Buttons are working for each appointment. I commented out some email code because it was causing things to break. When a doctor cancels an appointment, they delete it, but when a patient cancels an appointment, it gets added back into the appointments pool for the search page. Again, formatting could probably be changed to make things look better.