Branch of Gabriel's branch
Added edit button onto doctor's appointment page.
Created helpdoctor.php because we had inconsistencies in the page
options at the top of the page. "original help.php was being used for
doctor profile and was showing a medical history option that shouldn't
be visible for doctor." Updated color of delete button and options at
the top of the appointment page to be the same blue.